Alterei a entidade filme para condizer com as infos buscadas da api do tmdb

+++ src/main/java/com/filmes/avaliador/dto/request/AvaliacaoDTO.java
@@ -13,6 +13,6 @@ public record AvaliacaoDTO(
         @NotBlank(message = "Campo obrigatório")
         String idUsuario,
         @NotNull(message = "Campo obrigatório")
-        Integer idFilme
+        Long idFilme
 ) {
 }
